backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
Pillow>=10.0
jazzmin>=3.0.0
